Implemented on function in service worker context
Currently, on only functions for the schedule and cancel events.
In order to get this working, it was necessary to recreate a lot of code from the preexisting plugin.
This is because the plugin only dispatches events to the active webview using the cordova interface.
Since the service worker does not use cordova, the javascript interface had to be altered.
Also, the core functions that handle the native notification interface in the plugin were not exposed so those functions,
which could have easily been reused, had to be recreated.